Extend InMemoryEmfModel to get around Epsilon reusing resources
If we inherit from EmfModel, and we create two EMFViewsModel instances
like this:

  var a = new EMFViewsModel();
  a.setModelFileUri("a.eview");
  var b = new EMFViewsModel();
  b.setModelFileUri("a.eview");

then `a` will create and load a ViewResource instance, but `b`
will *reuse* that same instance, since they have the same URIs.  Even if
we dispose of `a` and trash it.  Even if we setCacheEnabled(false), or
setStoredOnDisposal(false), the resource is reused.

Switching to InMemoryEmfModel does not have this issue, since we are in
charge of loading the ViewResource ourselves.